% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrY_genes_data.R
\docType{data}
\name{default_chrY_genes}
\alias{default_chrY_genes}
\title{Default chromosome Y genes}
\format{
A character vector of ENSEMBL gene IDs
}
\usage{
default_chrY_genes
}
\description{
A vector of ENSEMBL gene IDs for genes located on chromosome Y
used for sex determination in RNA-seq analysis (PAR-Y genes are excluded from this list).
}
\keyword{datasets}
