% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_input_type}
\alias{detect_input_type}
\title{Detect input type automatically}
\usage{
detect_input_type(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
"counts" or "log"
}
\description{
Detect input type automatically
}
\keyword{internal}
