% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{.adjusted_rand_index}
\alias{.adjusted_rand_index}
\title{Adjusted Rand Index (ARI)
Compares clustering agreement between before and after correction}
\usage{
.adjusted_rand_index(pcs_before, pcs_after, n_clusters = NULL)
}
\description{
Adjusted Rand Index (ARI)
Compares clustering agreement between before and after correction
}
\keyword{internal}
