% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_mismatch.R
\name{.calculate_contamination_zone}
\alias{.calculate_contamination_zone}
\title{Calculate contamination zone parameters}
\usage{
.calculate_contamination_zone(
  data,
  contamination_slope,
  contamination_fraction
)
}
\description{
Calculate contamination zone parameters
}
\keyword{internal}
