% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_mismatch.R
\name{.classify_samples}
\alias{.classify_samples}
\title{Classify samples based on expression and contamination zone}
\usage{
.classify_samples(data, zone_params, threshold)
}
\description{
Classify samples based on expression and contamination zone
}
\keyword{internal}
