% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{.cluster_compactness_separation}
\alias{.cluster_compactness_separation}
\title{Cluster compactness and separation
Computes within-cluster and between-cluster distances}
\usage{
.cluster_compactness_separation(pcs, labels)
}
\description{
Cluster compactness and separation
Computes within-cluster and between-cluster distances
}
\keyword{internal}
