% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{.continuous_concordance}
\alias{.continuous_concordance}
\title{Continuous variable concordance}
\usage{
.continuous_concordance(Xlog_before, Xlog_after, cont_var)
}
\description{
Continuous variable concordance
}
\keyword{internal}
