% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{.create_comprehensive_summary}
\alias{.create_comprehensive_summary}
\title{Create comprehensive summary table with interpretations}
\usage{
.create_comprehensive_summary(
  assessments,
  run_names,
  knn_jaccard_vec,
  hvg_overlap_vec,
  biology_consistency_concordance,
  pvca_list,
  adjusted_rand_index_vec,
  ari_n_clusters_vec,
  cluster_compactness_separation_list,
  biology_type
)
}
\description{
Create comprehensive summary table with interpretations
}
\keyword{internal}
