% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{.create_consolidated_table}
\alias{.create_consolidated_table}
\title{Create consolidated table with all metrics in wide format
Structure: one row per metric, one column per run}
\usage{
.create_consolidated_table(
  assessments,
  run_names,
  knn_jaccard_vec,
  hvg_overlap_vec,
  biology_consistency_concordance,
  pvca_list,
  adjusted_rand_index_vec,
  ari_n_clusters_vec,
  cluster_compactness_separation_list
)
}
\description{
Create consolidated table with all metrics in wide format
Structure: one row per metric, one column per run
}
\keyword{internal}
