% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_mismatch.R
\name{.create_sex_mismatch_plot}
\alias{.create_sex_mismatch_plot}
\title{Create sex mismatch visualization}
\usage{
.create_sex_mismatch_plot(data, classifications, zone_params, anchors, title)
}
\description{
Create sex mismatch visualization
}
\keyword{internal}
