% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_combat.R
\name{.ebayes_shrink_location}
\alias{.ebayes_shrink_location}
\title{Empirical Bayes shrinkage for location parameters (gamma)}
\usage{
.ebayes_shrink_location(gamma_gb, n_anchors_gb, verbose = TRUE)
}
\description{
Empirical Bayes shrinkage for location parameters (gamma)
}
\keyword{internal}
