% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_combat.R
\name{.ebayes_shrink_scale}
\alias{.ebayes_shrink_scale}
\title{Empirical Bayes shrinkage for scale parameters (delta)}
\usage{
.ebayes_shrink_scale(delta_gb, n_anchors_gb, verbose = TRUE)
}
\description{
Empirical Bayes shrinkage for scale parameters (delta)
}
\keyword{internal}
