% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_mismatch.R
\name{.extract_sex_chromosome_expression}
\alias{.extract_sex_chromosome_expression}
\title{Extract and normalize XIST and Y chromosome expression}
\usage{
.extract_sex_chromosome_expression(
  counts_cpm,
  metadata,
  xist_gene,
  chry_genes,
  id_col,
  sex_col
)
}
\description{
Extract and normalize XIST and Y chromosome expression
}
\keyword{internal}
