% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{.factor_variance_preservation}
\alias{.factor_variance_preservation}
\title{Factor variance preservation
Assesses how well biological variance is preserved after correction}
\usage{
.factor_variance_preservation(pcs_before, pcs_after, biology)
}
\description{
Factor variance preservation
Assesses how well biological variance is preserved after correction
}
\keyword{internal}
