% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_detection.R
\name{.filter_low_genes}
\alias{.filter_low_genes}
\title{Filter genes with low counts}
\usage{
.filter_low_genes(mat, min_total, min_count, min_samples)
}
\description{
Filter genes with low counts
}
\keyword{internal}
