% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_detection.R
\name{.identify_outliers_2d}
\alias{.identify_outliers_2d}
\title{Identify outliers in 2D coordinates}
\usage{
.identify_outliers_2d(coords, n_sd)
}
\description{
Identify outliers in 2D coordinates
}
\keyword{internal}
