% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_detection.R
\name{.iterative_outlier_detection}
\alias{.iterative_outlier_detection}
\title{Iteratively detect and remove outliers}
\usage{
.iterative_outlier_detection(mat, n_sd, max_iterations, verbose)
}
\description{
Iteratively detect and remove outliers
}
\keyword{internal}
