% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{.multi_group_eta_squared_cor}
\alias{.multi_group_eta_squared_cor}
\title{Multi-group eta-squared correlation}
\usage{
.multi_group_eta_squared_cor(Xlog_before, Xlog_after, biology)
}
\description{
Multi-group eta-squared correlation
}
\keyword{internal}
