% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{.multi_group_fstat_cor}
\alias{.multi_group_fstat_cor}
\title{Multi-group F-statistic correlation}
\usage{
.multi_group_fstat_cor(Xlog_before, Xlog_after, biology)
}
\description{
Multi-group F-statistic correlation
}
\keyword{internal}
