% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.post_counts_guard}
\alias{.post_counts_guard}
\title{Post-process counts (guard against invalid values)}
\usage{
.post_counts_guard(x, clip_counts = TRUE, round_counts = TRUE)
}
\arguments{
\item{x}{numeric matrix}

\item{clip_counts}{logical; clip negative values to zero}

\item{round_counts}{logical; round to integers}
}
\value{
processed matrix
}
\description{
Post-process counts (guard against invalid values)
}
\keyword{internal}
