% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{.pvca}
\alias{.pvca}
\title{Principal Variance Component Analysis (PVCA)
Quantifies variance explained by batch vs biological factors}
\usage{
.pvca(pcs, batch, biology = NULL)
}
\description{
Principal Variance Component Analysis (PVCA)
Quantifies variance explained by batch vs biological factors
}
\keyword{internal}
