% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_mismatch.R
\name{.strip_gene_versions}
\alias{.strip_gene_versions}
\title{Strip gene version numbers from Ensembl IDs}
\usage{
.strip_gene_versions(counts)
}
\description{
Strip gene version numbers from Ensembl IDs
}
\keyword{internal}
