% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_detection.R
\name{.validate_outlier_inputs}
\alias{.validate_outlier_inputs}
\title{Validate inputs for outlier detection}
\usage{
.validate_outlier_inputs(
  data_matrix,
  n_sd,
  min_gene_count,
  min_gene_samples,
  max_iterations
)
}
\description{
Validate inputs for outlier detection
}
\keyword{internal}
