% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identify_anchors}
\alias{identify_anchors}
\title{Identify anchor samples and check coverage}
\usage{
identify_anchors(batch, sample_id, min_batches = 2)
}
\arguments{
\item{batch}{factor}

\item{sample_id}{factor}

\item{min_batches}{minimum number of batches a sample must appear in to be an anchor}
}
\value{
list with anchor_ids, anchor_indices, coverage_report
}
\description{
Identify anchor samples and check coverage
}
\keyword{internal}
