% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{print.anchor_assessment_merged}
\alias{print.anchor_assessment_merged}
\title{Print method for merged assessments}
\usage{
\method{print}{anchor_assessment_merged}(x, ...)
}
\arguments{
\item{x}{Object to print.}
\item{...}{Additional arguments passed to methods.}
}
\description{
Print method for merged assessments.
}
