% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_fit_apply.R
\name{print.anchor_ridge_fit}
\alias{print.anchor_ridge_fit}
\title{Print method for anchor_ridge_fit}
\usage{
\method{print}{anchor_ridge_fit}(x, ...)
}
\arguments{
\item{x}{Object to print.}
\item{...}{Additional arguments passed to methods.}
}
\description{
Print method for anchor_ridge_fit
}
