% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_ref_batch}
\alias{validate_ref_batch}
\title{Validate reference batch}
\usage{
validate_ref_batch(ref_batch, batch, sample_id)
}
\arguments{
\item{ref_batch}{character or NULL}

\item{batch}{factor}

\item{sample_id}{factor}
}
\value{
validated ref_batch (character) or NULL
}
\description{
Checks that ref_batch exists in the batch factor and issues warnings
about anchor coverage.
}
\keyword{internal}
