% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_correction.R
\name{print.anchor_assessment}
\alias{print.anchor_assessment}
\title{Print method for anchor_assessment}
\usage{
\method{print}{anchor_assessment}(x, ...)
}
\arguments{
\item{x}{Object to print.}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for anchor_assessment
}
